edit

+++ app/Http/Controllers/Admin/PostController.php
@@ -6,11 +6,22 @@ use App\Models\Tag;
 use App\Models\Post;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
+    protected $validation_rules = [
+        'title' => 'required|string|max:100',
+        'slug' => ['required','string','max:100',],
+        'category_id' => 'required|exists:categories,id',
+        'tags' => 'nullable|array',
+        'tags.*' =>'integer|exists:tags,id',
+        'image' =>'required_without:content|nullable|url',
+        'content' =>'required_without:image|nullable|string|max:5000',
+        'excerpt' => 'nullable|string|max:200'
+    ];
     /**
      * Display a listing of the resource.
      *
@@ -20,6 +31,12 @@ class PostController extends Controller
     {
         $posts = Post::paginate(5);
 
+        return view('admin.posts.index', compact('posts'));
+    }
+
+    public function myIndex()
+    {
+        $posts = Auth::user()->posts()->paginate(5);
         return view('admin.posts.index', compact('posts'));
     }
 
@@ -47,16 +64,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title' => 'required|string|max:100',
-            'slug' => 'required|string|max:100|unique:posts',
-            'category_id' => 'required|exists:categories,id',
-            'tags' => 'nullable|array',
-            'tags.*' =>'integer|exists:tags,id',
-            'image' =>'required_without:content|nullable|url',
-            'content' =>'required_without:image|nullable|string|max:5000',
-            'excerpt' => 'nullable|string|max:200'
-        ]);
+        $this->validation_rules['slug'][] = 'unique:posts';
+        $request->validate($this->validation_rules);
 
         // dump($request->all());
         $formData = $request->all() + [
@@ -66,7 +75,7 @@ class PostController extends Controller
         $post = Post::create($formData); // con protected $fillable nel model
         $post->tags()->sync($formData['tags']);
 
-        return redirect()->route('admin.posts.index');
+        return redirect()->route('admin.posts.show', ['post' => $post]);
     }
 
     /**
@@ -88,7 +97,16 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        return view('admin.posts.edit', compact('post'));
+        if (auth::id() != $post->user_id) abort(401);
+
+        $categories = Category::all();
+        $tags = Tag::all();
+
+        return view('admin.posts.edit', [
+            'post' => $post,
+            'categories' => $categories,
+            'tags' => $tags
+        ]);
     }
 
     /**
@@ -100,11 +118,14 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $this->validation_rules['slug'][] = Rule::unique('posts')->ignore($post->id);
+        $request->validate($this->validation_rules);
+
         $formData = $request->all();
-
         $post->update($formData);// con protected $fillable nel model
+        $post->tags()->sync($formData['tags']);
 
-        return redirect()->route('posts.show', ['post' => $post]);
+        return redirect()->route('admin.posts.show', ['post' => $post]);
     }
 
     /**
@@ -115,6 +136,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (auth::id() != $post->user_id) abort(401);
+
         $post->tags()->detach();
         $post->delete();
